+++ examples/getobject.go
@@ -6,11 +6,12 @@ import (
 )
 
 func main() {
-	// create api object
-	a, _ := goidoit.NewApi("http://example.com/src/jsonrpc.php", "yourapikey")
+	// Debug and SSL Skip
+	//goidoit.Debug(true)
+	//goidoit.SkipTLSVerify(true)
 
-	// enable debug
-	// goidoit.Debug(true)
+	// create api object using api url and your api key
+	a, _ := goidoit.NewApi("https://example.com/src/jsonrpc.php", "yourapikey")
 
 	// select via string
 	viaString, _ := a.GetObject("Root location")
