+++ examples/getreport.go
@@ -6,15 +6,17 @@ import (
 )
 
 func main() {
-	// create api object
-	a, _ := goidoit.NewApi("https://example.com/src/jsonrpc.php", "yourapikey")
-
-	// enable debug
+	// Debug and SSL Skip
 	//goidoit.Debug(true)
+	//goidoit.SkipTLSVerify(true)
+
+	// create api object using api url and your api key
+	a, _ := goidoit.NewApi("https://example.com/src/jsonrpc.php", "yourapikey")
 
 	data, _ := a.GetReport(1)
 
 	for _, v := range data.Result {
+		fmt.Println(v)
 		/*
 			do sth with you report data
 		*/
