+++ internals_test.go
@@ -1,0 +1,71 @@
+/*
+	Go library for simple i-doit api usage
+
+	Copyright (C) 2017 Carsten Seeger
+
+	This program is free software: you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version.
+
+	This program is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+	@author Carsten Seeger
+	@copyright Copyright (C) 2017 Carsten Seeger
+	@license http://www.gnu.org/licenses/gpl-3.0 GNU General Public License 3
+	@link https://github.com/cseeger-epages/i-doit-go-api
+*/
+
+package goidoit
+
+import (
+	"testing"
+)
+
+type inttest struct {
+	n        int
+	expected int
+}
+
+type constructortest struct {
+	input    []string
+	expected bool
+}
+
+func TestGetID(t *testing.T) {
+	var tests = []inttest{
+		{1, 1},
+		{5, 5},
+		{100, 100},
+	}
+	for _, pairs := range tests {
+		var test = 0
+		TResetID()
+		for i := 1; i <= pairs.n; i++ {
+			test = TgetID()
+		}
+		if test != pairs.expected {
+			t.Errorf("Expected %v got %v\n", pairs.expected, test)
+		}
+	}
+}
+
+func TestDebugPrint(t *testing.T) {
+}
+
+func TestDebug(t *testing.T) {
+	Debug(false)
+	if TgetDebug() {
+		t.Errorf("Expected false got true\n")
+	}
+	Debug(true)
+	if !TgetDebug() {
+		t.Errorf("Expected true got false\n")
+	}
+}
